Move Toolbox component in toolbox module
Seems to make more sense. I think the _Toolbox class can go, though.
