+++ gaphor/ui/mainwindow.py
@@ -5,13 +5,13 @@ The main application window.
 import logging
 import os.path
 from zope import component
+from zope.interface import implementer
 
 import pkg_resources
 from gi.repository import GLib
 from gi.repository import Gdk
 from gi.repository import GdkPixbuf
 from gi.repository import Gtk
-from zope.interface import implementer
 
 from gaphor import UML
 from gaphor.UML.event import ModelFactoryEvent
@@ -28,13 +28,11 @@ from gaphor.services.filemanager import FileManagerStateChanged
 from gaphor.services.undomanager import UndoManagerStateChanged
 from gaphor.ui.accelmap import load_accel_map, save_accel_map
 from gaphor.ui.diagrampage import DiagramPage
-from gaphor.ui.diagramtoolbox import TOOLBOX_ACTIONS
 from gaphor.ui.event import DiagramPageChange, Diagram
-from gaphor.ui.interfaces import IDiagramPageChange
 from gaphor.ui.interfaces import IUIComponent
 from gaphor.ui.layout import deserialize
 from gaphor.ui.namespace import Namespace
-from gaphor.ui.toolbox import Toolbox as _Toolbox
+from gaphor.ui.toolbox import Toolbox
 
 log = logging.getLogger(__name__)
 
@@ -393,105 +391,6 @@ class MainWindow(object):
 
 
 Gtk.AccelMap.add_filter("gaphor")
-
-
-@implementer(IUIComponent, IActionProvider)
-class Toolbox(object):
-
-    title = _("Toolbox")
-    placement = ("left", "diagrams")
-
-    component_registry = inject("component_registry")
-    main_window = inject("main_window")
-    properties = inject("properties")
-
-    menu_xml = """
-      <ui>
-        <menubar name="mainwindow">
-          <menu action="diagram">
-            <separator/>
-            <menuitem action="reset-tool-after-create" />
-            <separator/>
-          </menu>
-        </menubar>
-      </ui>
-    """
-
-    def __init__(self):
-        self._toolbox = None
-        self.action_group = build_action_group(self)
-        self.action_group.get_action("reset-tool-after-create").set_active(
-            self.properties.get("reset-tool-after-create", True)
-        )
-
-    def open(self):
-        widget = self.construct()
-        self.main_window.window.connect_after(
-            "key-press-event", self._on_key_press_event
-        )
-        return widget
-
-    def close(self):
-        if self._toolbox:
-            self._toolbox.destroy()
-            self._toolbox = None
-
-    def construct(self):
-        toolbox = _Toolbox(TOOLBOX_ACTIONS)
-        toolbox.show()
-
-        toolbox.connect("destroy", self._on_toolbox_destroyed)
-        self._toolbox = toolbox
-        return toolbox
-
-    def _on_key_press_event(self, view, event):
-        """
-        Grab top level window events and select the appropriate tool based on the event.
-        """
-        if event.get_state() & Gdk.ModifierType.SHIFT_MASK or (
-            event.get_state() == 0 or event.get_state() & Gdk.ModifierType.MOD2_MASK
-        ):
-            keyval = Gdk.keyval_name(event.keyval)
-            self.set_active_tool(shortcut=keyval)
-
-    def _on_toolbox_destroyed(self, widget):
-        self._toolbox = None
-
-    @toggle_action(name="reset-tool-after-create", label=_("_Reset tool"), active=False)
-    def reset_tool_after_create(self, active):
-        self.properties.set("reset-tool-after-create", active)
-
-    @component.adapter(IDiagramPageChange)
-    def _on_diagram_page_change(self, event):
-        self.update_toolbox(event.diagram_page.toolbox.action_group)
-
-    def update_toolbox(self, action_group):
-        """
-        Update the buttons in the toolbox. Each button should be connected
-        by an action. Each button is assigned a special _action_name_
-        attribute that can be used to fetch the action from the ui manager.
-        """
-        if not self._toolbox:
-            return
-
-        for button in self._toolbox.buttons:
-
-            action_name = button.action_name
-            action = action_group.get_action(action_name)
-            if action:
-                button.set_related_action(action)
-
-    def set_active_tool(self, action_name=None, shortcut=None):
-        """
-        Set the tool based on the name of the action
-        """
-        # HACK:
-        toolbox = self._toolbox
-        if shortcut and toolbox:
-            action_name = toolbox.shortcuts.get(shortcut)
-            log.debug("Action for shortcut %s: %s" % (shortcut, action_name))
-            if not action_name:
-                return
 
 
 @implementer(IUIComponent, IActionProvider)
